# Improve `sanitize_sql_like` performance